Organised project directory into contained packages.
