+++ src/main/Entry.java
@@ -1,4 +1,4 @@
-package sys;
+package main;
 
 public class Entry {
     public static void main(String[] args) throws Exception {
